+++ UVtools.Core/Operations/OperationLayerReHeight.cs
@@ -119,5 +119,12 @@ namespace UVtools.Core.Operations
                 return (IsMultiply ? 'x' : '÷') + $" {Modifier} → {LayerCount} layers at {LayerHeight}mm";
             }
         }
+
+        public override string ToString()
+        {
+            var result = $"[Layer Count: {Item.LayerCount}] [Layer Height: {Item.LayerHeight}]" + LayerRangeString;
+            if (!string.IsNullOrEmpty(ProfileName)) result = $"{ProfileName}: {result}";
+            return result;
+        }
     }
 }